Added folder of new changes, and basic functionalities
This will serve as the editing basis. Please add changes on TOP of this,
but remember to save a backup copy if things go awry.
What I did here: I made the 4 webpages. Connected the webpages up. Added
icons that link FROM the homepage to:
a) The Record Page
b) The View Page
c) The Settings Page

Note:
- I have NOT yet added an icon/button to go back to the home page. Will
do so later.